rearrange for readable diff
